Consistency: instrumentation classes should be in files ending with Instrumentation.java
This is so we can direct load requests to a special class-loader that can be unloaded after use.
